fix: 인원 수 자료형 int로 변경

+++ src/main/java/com/example/haemo/haemo/Data/Post.java
@@ -28,7 +28,7 @@ public class Post {
     private String nickname;
 
     @Column(name = "person")
-    private Long person;
+    private int person;
 
     @Column(name = "category")
     private String category;
